Neatened up platformio file
Removed the duplicate entries and put them into global [env]. Debug
tool remains part of each one, just in case of a chip that uses a
different debugger.

Also removed the library include. It's not necessary (and even if
it were, it's better to add the ones you need specifically).