Discovered interesting bug while testing PCB
Implemented a boards file for the custom PCB's that use the ESP32-S3
with 4MB/2MB of Flash/PSRAM, and therefore require different settings
to the devkitC. I copied most of them from the Adafruit feather, except
the PSRAM compile flag (because it would immediately crash).

However, I discovered some weird behaviour, which also happens on the
devkitC as it turns out. Any time spent 'paused' using the JTAG
debugger e.g. at a breakpoint or stepping through, is counted against
the next call to delay()! It makes no sense, but unfortunately the
functions are written in assembly, making it hard to figure out...